backend/requirements.txt
firebase-admin
pandas
numpy
python-dotenv
google-generativeai
